feat question error handle refactor database int to bigint

+++ src/main/java/com/sundingyi/forum/controller/PublishController.java
@@ -31,7 +31,7 @@ public class PublishController {
     public String doPublish(@RequestParam("title") String title,
                             @RequestParam("description") String description,
                             @RequestParam("tag") String tag,
-                            @RequestParam(name = "id", required = false) Integer id,
+                            @RequestParam(name = "id", required = false) Long id,
                             HttpServletRequest httpServletRequest,
                             Model model) {
         model.addAttribute("title", title);
@@ -58,7 +58,7 @@ public class PublishController {
     }
     
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable("id") Integer id,
+    public String edit(@PathVariable("id") Long id,
                        Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
